+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for Python 3.14
+- CLI argument to define path to database when creating config
 
 ## [0.6.0] - 2025-09-14